+++ src/collections/views/CollectionDetails.tsx
@@ -18,8 +18,9 @@ import useNotifier from "@saleor/hooks/useNotifier";
 import usePaginator, {
   createPaginationState
 } from "@saleor/hooks/usePaginator";
-import { commonMessages } from "@saleor/intl";
+import { commonMessages, errorMessages } from "@saleor/intl";
 import useProductSearch from "@saleor/searches/useProductSearch";
+import { arrayDiff } from "@saleor/utils/arrays";
 import createDialogActionHandlers from "@saleor/utils/handlers/dialogActionHandlers";
 import createMetadataUpdateHandler from "@saleor/utils/handlers/metadataUpdateHandler";
 import { mapEdgesToItems } from "@saleor/utils/maps";
@@ -28,7 +29,6 @@ import {
   usePrivateMetadataUpdate
 } from "@saleor/utils/metadata/updateMetadata";
 import { getParsedDataForJsonStringField } from "@saleor/utils/richText/misc";
-import { diff } from "fast-array-diff";
 import React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
@@ -100,7 +100,8 @@ export const CollectionDetails: React.FC<CollectionDetailsProps> = ({
       if (backgroundImageError) {
         notify({
           status: "error",
-          text: intl.formatMessage(commonMessages.somethingWentWrong)
+          title: intl.formatMessage(errorMessages.imgageUploadErrorTitle),
+          text: intl.formatMessage(errorMessages.imageUploadErrorText)
         });
       }
     }
@@ -214,11 +215,14 @@ export const CollectionDetails: React.FC<CollectionDetailsProps> = ({
               input
             }
           });
-          const diffChannels = diff(
-            collectionChannelsChoices,
-            formData.channelListings,
-            (a, b) => a.id === b.id
+          const initialIds = collectionChannelsChoices.map(
+            channel => channel.id
           );
+          const modifiedIds = formData.channelListings.map(
+            channel => channel.id
+          );
+
+          const idsDiff = arrayDiff(initialIds, modifiedIds);
 
           updateChannels({
             variables: {
@@ -229,10 +233,7 @@ export const CollectionDetails: React.FC<CollectionDetailsProps> = ({
                   isPublished: channel.isPublished,
                   publicationDate: channel.publicationDate
                 })),
-                removeChannels:
-                  diffChannels.removed?.map(
-                    removedChannel => removedChannel.id
-                  ) || []
+                removeChannels: idsDiff.removed
               }
             }
           });