+++ src/fragments/types/SelectedVariantAttributeFragment.ts
@@ -32,6 +32,8 @@ export interface SelectedVariantAttributeFragment_attribute_choices_edges_node {
   reference: string | null;
   richText: any | null;
   boolean: boolean | null;
+  date: any | null;
+  dateTime: any | null;
 }
 
 export interface SelectedVariantAttributeFragment_attribute_choices_edges {
@@ -73,6 +75,8 @@ export interface SelectedVariantAttributeFragment_values {
   reference: string | null;
   richText: any | null;
   boolean: boolean | null;
+  date: any | null;
+  dateTime: any | null;
 }
 
 export interface SelectedVariantAttributeFragment {