+++ src/intl.ts
@@ -52,6 +52,7 @@ export const commonMessages = defineMessages({
   generalInformations: {
     defaultMessage: "General Information"
   },
+
   insufficientPermissions: {
     defaultMessage: "Insufficient permissions"
   },
@@ -107,6 +108,23 @@ export const commonMessages = defineMessages({
   },
   yes: {
     defaultMessage: "Yes"
+  },
+  date: {
+    defaultMessage: "Date"
+  },
+  time: {
+    defaultMessage: "Time",
+    description: "independent of any particular day, eg. 11:35"
+  }
+});
+
+export const errorMessages = defineMessages({
+  imgageUploadErrorTitle: {
+    defaultMessage: "Couldn't process image"
+  },
+  imageUploadErrorText: {
+    defaultMessage:
+      "There was a poblem with the file you uploaded as an image and it couldn't be used. Please try a different file."
   }
 });