+++ src/products/urls.ts
@@ -67,6 +67,7 @@ export const productListUrl = (params?: ProductListUrlQueryParams): string =>
 
 export const productPath = (id: string) => urlJoin(productSection + id);
 export type ProductUrlDialog =
+  | "add-variants"
   | "remove"
   | "remove-variants"
   | "assign-attribute-value"