+++ src/types/globalTypes.ts
@@ -89,6 +89,8 @@ export enum AttributeErrorCode {
 
 export enum AttributeInputTypeEnum {
   BOOLEAN = "BOOLEAN",
+  DATE = "DATE",
+  DATE_TIME = "DATE_TIME",
   DROPDOWN = "DROPDOWN",
   FILE = "FILE",
   MULTISELECT = "MULTISELECT",
@@ -1130,6 +1132,10 @@ export interface AttributeInput {
   slug: string;
   values?: (string | null)[] | null;
   valuesRange?: IntRangeInput | null;
+  dateTimeRange?: DateTimeRangeInput | null;
+  dateRange?: DateRangeInput | null;
+  dateTime?: any | null;
+  date?: any | null;
   boolean?: boolean | null;
 }
 
@@ -1167,6 +1173,8 @@ export interface AttributeValueInput {
   references?: string[] | null;
   richText?: any | null;
   boolean?: boolean | null;
+  date?: any | null;
+  dateTime?: any | null;
 }
 
 export interface AttributeValueTranslationInput {
@@ -1176,7 +1184,8 @@ export interface AttributeValueTranslationInput {
 
 export interface BulkAttributeValueInput {
   id?: string | null;
-  values: string[];
+  values?: (string | null)[] | null;
+  boolean?: boolean | null;
 }
 
 export interface CatalogueInput {